[TS/TC][87321900] WIP: default value
An incomplete default value implementation. Obviously, `set_values` is
not the place to be doing this. It's not AttributeSet's responsibility
to check for the existence of a default value. It should be, at worst,
in Attribute, and at best a message in an object that obviates the need
for a conditional. Just give me the value.

Tabling this for now to spike out class attributes and instance
attributes as different objects. A class attribute should not have a
value. We're not supporting class-scope attribute values, but the code
would lead one to believe we are. If this spike works out, the default
value implemenation should follow.