[TS/TC][91426904] Add custom coercer params
We did not solve this problem by adding a mapper config object like
expected. Instead, we added an option to the attribute DSL that supports
coercer object initialization arguments. This allows us to inject data
into coercer objects on a per-attribute basis and filter values from the
data source in any way.